+++ app/guests/reports/scores.py
@@ -10,6 +10,8 @@ from typing import Any
 from mysql.connector.connection import MySQLConnection
 from mysql.connector.pooling import PooledMySQLConnection
 
+from app.shows.reports.show_details import retrieve_show_date_by_id
+
 
 def retrieve_scoring_exceptions(
     guest_id: int, database_connection: MySQLConnection | PooledMySQLConnection
@@ -208,3 +210,54 @@ def retrieve_all_three_pointers(
         )
 
     return three_pointers
+
+
+def retrieve_all_missing_scores(
+    database_connection: MySQLConnection | PooledMySQLConnection,
+) -> list[dict[str, str | int | bool | None]]:
+    """Retrieve all Not My Job guests with no scores entered."""
+    if not database_connection.is_connected():
+        database_connection.reconnect()
+
+    query = """
+        SELECT s.showdate, s.bestof, s.repeatshowid, g.guest,
+        g.guestslug, gm.guestscore, gm.exception
+        FROM ww_showguestmap gm
+        JOIN ww_guests g ON g.guestid = gm.guestid
+        JOIN ww_shows s ON s.showid = gm.showid
+        WHERE gm.guestscore IS NULL
+        AND g.guestslug <> 'none'
+        AND s.showdate < NOW()
+        ORDER BY s.showdate ASC;
+    """
+    cursor = database_connection.cursor(dictionary=True)
+    cursor.execute(query)
+    results = cursor.fetchall()
+    cursor.close()
+
+    if not results:
+        return None
+
+    _guests = []
+    for row in results:
+        _guests.append(
+            {
+                "date": row["showdate"],
+                "best_of": bool(row["bestof"]),
+                "repeat": bool(row["repeatshowid"]),
+                "original_show_date": (
+                    retrieve_show_date_by_id(
+                        show_id=row["repeatshowid"],
+                        database_connection=database_connection,
+                    )
+                    if row["repeatshowid"]
+                    else None
+                ),
+                "name": row["guest"],
+                "slug": row["guestslug"],
+                "score": row["guestscore"],
+                "exception": bool(row["exception"]),
+            }
+        )
+
+    return _guests